Update map with 3 difficulty: Easy, Medium, Hard without having object

+++ fileh/Level.h
@@ -8,7 +8,7 @@
 
 class Level : public State {
 private:
-	float difficulty;
+	int difficulty;
 
 	float deltaSpeed;
